+++ app/src/main/java/org/sopt/hyundaicarddive/HyunDaiCardDiveApplication.kt
@@ -1,0 +1,8 @@
+package org.sopt.hyundaicarddive
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class HyunDaiCardDiveApplication: Application() {
+}
