+++ app/src/main/java/org/sopt/hyundaicarddive/core/component/DummyComponent.kt
@@ -1,0 +1,2 @@
+package org.sopt.hyundaicarddive.core.component
+
