+++ app/src/main/java/org/sopt/hyundaicarddive/core/navigation/RouteModel.kt
@@ -1,0 +1,2 @@
+package org.sopt.hyundaicarddive.core.navigation
+
