+++ app/src/main/java/org/sopt/hyundaicarddive/data/mapper/todata/DummyRequestMapper.kt
@@ -1,0 +1,2 @@
+package org.sopt.hyundaicarddive.data.mapper.todata
+
