+++ app/src/main/java/org/sopt/hyundaicarddive/domain/model/DummyModel.kt
@@ -1,0 +1,4 @@
+package org.sopt.hyundaicarddive.domain.model
+
+class DummyModel {
+}
