+++ app/src/main/java/org/sopt/hyundaicarddive/domain/repository/DummyRepository.kt
@@ -1,0 +1,4 @@
+package org.sopt.hyundaicarddive.domain.repository
+
+class DummyRepository {
+}
