+++ app/src/main/java/org/sopt/hyundaicarddive/domain/usecase/DummyUseCase.kt
@@ -1,0 +1,4 @@
+package org.sopt.hyundaicarddive.domain.usecase
+
+class DummyUseCase {
+}
