+++ app/src/main/java/org/sopt/hyundaicarddive/presentation/model/DummyModel.kt
@@ -1,0 +1,4 @@
+package org.sopt.hyundaicarddive.presentation.model
+
+class DummyModel {
+}
