+++ app/src/main/java/org/sopt/hyundaicarddive/presentation/ui/main/MainActivity.kt
@@ -1,0 +1,20 @@
+package org.sopt.hyundaicarddive.presentation.ui.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import dagger.hilt.android.AndroidEntryPoint
+import org.sopt.hyundaicarddive.ui.theme.HYUNDAICARDDIVETheme
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            HYUNDAICARDDIVETheme {
+            }
+        }
+    }
+}
